Adds function relpath
This is needed later to calculate the relative path for the Index Page
to the different API pages. Issue #43

I thought maybe smaller PR are easier to merge.

* I had to redo it: the last one had some problems in certain
situations.
* I wrote an test case using results generated with python's relpath
(expecting that this should work ok)

helper function: findlast: I know one could write it a bit shorter with
a for loop - but in my tests such while loops are about 10% faster and I
keep it for other things too.
If you prefer it different no problem if you want to rewrite it.

This time I use for the tests the: Base.path_separator because the
previous failed on windows.